+++ modules/theme-tailwind-magento/magento/cms/blocks/pb-products.vue
@@ -6,18 +6,19 @@
   </div>
 
   <div v-if="pbProducts.type === 'grid'">
-    <ProductGrid :products="products" />
+    <ProductGrid v-if="products.length" :products="products" />
   </div>
 </template>
+
 <script setup lang="ts">
 import useGetProductsByIds from '#ioc/services/useGetProductsByIds'
-import ProductGrid from '#ioc/molecules/ProductGrid'
 import usePbProducts from '#ioc/composables/usePbProducts'
 import useResource from '#ioc/composables/useResource'
 import hydrateWhenVisible from '#ioc/utils/hydration/hydrateWhenVisible'
 import { computed } from 'vue'
 
 const ProductCarousel = hydrateWhenVisible(() => import('#ioc/organisms/ProductCarousel'))
+const ProductGrid = hydrateWhenVisible(() => import('#ioc/organisms/ProductGrid'))
 
 const props = defineProps({ el: { type: Object, default: null } })
 
@@ -28,8 +29,10 @@ const [data] = await useResource(
   () => pbProducts.skus,
   (skus) => getProductsByIds(skus),
 )
+
 const products = computed(() => data.value.products)
 </script>
+
 <script lang="ts">
 import { defineComponent } from 'vue'
