+++ modules/theme-tailwind/organisms/ProductListing.vue
@@ -35,7 +35,7 @@ import useI18n from '#ioc/composables/useI18n'
 import useRoute from '#ioc/composables/useRoute'
 import ToProduct from '#ioc/mappers/ToProduct'
 import ProductSort from '#ioc/molecules/ProductSort'
-import ProductGrid from '#ioc/molecules/ProductGrid'
+import ProductGrid from '#ioc/organisms/ProductGrid'
 import { computed, PropType } from 'vue'
 import hydrateWhenVisible from '#ioc/utils/hydration/hydrateWhenVisible'
 import hydrateWhenIdle from '#ioc/utils/hydration/hydrateWhenIdle'
